Dynamic content default value for value types
The default value will no longer be forced for properties that are value types, as it might incorrectly overwrite a valid value. For example, if the default value of a bool property is true, but a developer has set that property to false, the final value would be true. Same goes for integers: if the default value of a property is 100, but a developer has set it to 0, the final value would be 100.