Template module fixes and changes (#53)
* Dynamic content default value for value types

The default value will no longer be forced for properties that are value types, as it might incorrectly overwrite a valid value. For example, if the default value of a bool property is true, but a developer has set that property to false, the final value would be true. Same goes for integers: if the default value of a property is 100, but a developer has set it to 0, the final value would be 100.

* Open dynamic content blocks on double click

Dynamic content blocks can now be opened by double clicking on a row in the grid.

* Added "Redirect URL" setting

Added "Redirect URL" setting to the "login required" settings.