Updated import module entity properties (#91)
* Updated import module entity properties

Entity properties are now selected base on their order in Wiser instead of always alphabetically. The list is still ordered alphabetically though. Also refactored the 'get-entity-properties' call to use an endpoint of the Imports module instead of using the old query in the templates service.

* Updated entity property model usage

Removed custom EntityPropertyModel class from the Imports module and updated code to use the EntityPropertyModel that is available in the EntityProperties module.