Template connected users with Pusher (#127)
* Started work on listing connected users for templates.

* Added event that gets raised when connected users state changes.

* Completed templates connected users functionality

Users are now listed per unique connection instead of per username.

* Updated to use Pusher instead of SignalR

* Updated package-lock (to remove SignalR dependency)

* Fixed users not disconnecting properly

Fixed users staying in the connected users list even if they already closed the module or browser. Also removed a piece of code that belonged to SignalR that was accidentally left in the Startup code.

* Removed model

Removed a model that is no longer being used.