Routine, view and trigger templates development
Updated buttons and titles for the confirmation dialog when trying to rename or delete a routine, view, or trigger template. When renaming a trigger the old trigger will now get properly deleted. Loading routines will no longer generate an error due to trying to rename virtual templates. The necessary columns for triggers in wiser_template are now added to the CreateTables SQL script.