View, routine and trigger templates (#165)
* Define views and routines as templates

Allow views to be created through Wiser. This will also load all views that weren't created through Wiser so they can be saved in Wiser. Routines could already be created through Wiser, but this update will also show routines that weren't created in Wiser so they can be managed through Wiser.

* Finished up routines and views and added triggers

* Routine, view and trigger templates development

Updated buttons and titles for the confirmation dialog when trying to rename or delete a routine, view, or trigger template. When renaming a trigger the old trigger will now get properly deleted. Loading routines will no longer generate an error due to trying to rename virtual templates. The necessary columns for triggers in wiser_template are now added to the CreateTables SQL script.

* History tab now works for trigger templates

Some elements of the templates are no longer available for views, routines, and triggers (such as the published environments).