Pusher changes for template connected users (#204)
Changed the way Pusher works so the channel names and event names can
actually be used the way they were intended to be used. This allows the
`TemplateConnectedUsers` scripting to create a channel specific for the
current sub-domain. Also changes some scripting in `TaskAlerts` and
`Fields` so they don't break due to the Pusher changes.