Created TagHelper for loading Webpack scripts
This was needed because, for some reason, webpack generates more scripts on development than on production and those scripts are needed on development, otherwise it won't run. So I created this TagHelper which checks if each script exists in the Webpack manifest and only loads that script if it does.