Added new post message actions (#226)
* Added new post message actions

Two new post message actions have been added:
* `OpenItem`: Will open a Wiser item with the specified data.
* `GetAccessToken`: Will respond to the source window with the access
token. Will refresh if needed. Note: This will only work on test
environments, and for iframes that are loaded in the same main domain.

* Updated syntax

Early exit in the switch statement now uses `break` instead of `return`.
The case for `GetAccessToken` now exits early when the validation fails
to reduce indentation in the code.