Always parse value to string (#273)
When the value is set directly instead of by a replacement it can be an integer causing an exception to be thrown on the replace function.

By always parsing it to a string the replace function can be called. If the value was already a string it did already return as a string.