Changes how the templates module works with versions.
From now on, we will no longer create a new version for every update. When you save a template or component, it will now overwrite the latest version (as long as it's not published to live). When committing a template or component, that will automatically create a new version to work in from then on.