Use writing connection when getting latest version number.
The reading connection was used. If a new version was added in a transaction this function would return the old version.

This went wrong when committing a template directly to live a new version was created for live and then again. Now the writing connection is used the transaction is recognized when getting the latest version.