Fixed grid row delete button clipped off
When the command column in the grid is too narrow, the delete button gets clipped and ellipses are placed. However, the icon remains making it look broken.
This change widens the window, giving enough room for each column, and removes the text from the delete button to have it use less room.