Prevent lower versions from being published in the version control.
If a commit contains multiple templates/dynamic contents from which a template was published to a higher environment in another commit it would be published to the lowe version if the first commit is published.

This checks validates if the template is allowed to be published to the environment to ensure it can only go up.