Feature/version control improvements (#541)
* Create new version when commiting to development

* Fix grouping of dynamic components to show all commits

A dynamic component was only shown in one open commit. By adding the commit ID to the grouping this is now shown correctly.

* Set correct ID for external files on branch

The ID of the row of the template can be different on a branch. The ID of the main database was used causing mismatches. Based on template ID and version the correct row ID in the branch database is found.