Fix missing versions when using dropdown to deploy
When a template or dynamic content was deployed using the dropdowns no new version was created. When committing to the same environment the check to create a new version would fail because the latest version was already on that environment.

This check is now always executed for templates, instead of only for live. For dynamic contents this has been added since it was missing in the first place.