Feature/branch improve id mapping (#648)
* Add option to create item also on main branch

Add front-end checkbox when creating new item to tell the API to create is also on the main branch to ensure matching IDs.

* Create item in both databases

Initial setup to create an item in both databases.

* Create item in both databases.

* Ensure correct history when creating in both databases

Prevent create history of the item in the branch to prevent it being created a second time during a merge. Ensure published environment is always updated after the creation to trigger this in the merge.

* Add locks to prevent race conditions

* Map ID of newly created item

Map the ID to indicate that this item has already been mapped with the main database.