Fixed some parent ID issues (#851)
A combobox that saves its value as a link on a link type that has
"UseItemParentId" enabled would not retrieve its value properly. This is
now fixed.

In addition, setting "currentItemIsDestinationId" to true for a
multiselect in conjunction with "UseItemParentId" is now an invalid
combination and will show an error.

Co-authored-by: Hans <35447467+HansJuice@users.noreply.github.com>